Create Email-to-DomainWhoIsDataExtract.js
This code logs into a database, searches a table for email addresses that are not from popular providers like Gmail, Yahoo, AOL, and Outlook. It removes the username part of the email, leaving only the domain name. It then performs basic "whois" queries to gather usable information. This code is a proof of concept (PoC) and is surprisingly effective for large email lists, although it can be slow. The slowness issue can be easily addressed.